Refactor #210 PR0: trekk ut useChatData og useChatScroll fra Chat.tsx
Ren refactor — ingen atferdsendring. Chat.tsx (1462 linjer) splittet
i tre: felles typer (types.ts), datalag (hooks/useChatData.ts) og
scroll-lag (hooks/useChatScroll.ts). Alle eslint-disable-kommentarer
og subtile kommentarer er bevart ordrett. Re-eksport av ChatScope og
ChatMelding holder eksisterende callsites uendret.
