+++ lib/versjon.json
@@ -1,4 +1,4 @@
 {
-  "nummer": 34,
-  "versjon": "V3.1.34"
+  "nummer": 35,
+  "versjon": "V3.1.35"
 }
